requirements.txt.txt
pymongo==4.10.1
pandas==2.2.3
geopandas==1.0.1
shapely==2.0.6
matplotlib==3.9.2
contextily==1.6.2
geopy==2.4.1
pillow==11.0.0
numpy==2.1.3